I don't understand what this select would be for
my thought process:
this select doesn't run in a loop so its only going to evaluate once.
The first branch will take a single item from the buffered channel, but
it is discarded.
If the channel is already empty then default will be called, closing the
channel. But this doesn't get run again so I don't know what the goal
would be of the channel read case