fix desync on fast flicks out of canvas
this was caused by a difference in what the drawer and receiver drew.
The drawer would draw a line between the previous point to the end
point. The receiver only received the end point on end messages and so
the line it drew was just a point at the end. Now the end message
includes start and end payloads like the main draw event so the receiver
can draw the line between and all the way up to the end point